Return empty array for `parents` and `children` if receives malformed data.
Sometimes, MWS webservice fails and returns a http request (with status of 200)
and some of the data is malformed.

This commit returns an empty array for `children` and `parents` when the data is
malformed. Malformed data returns an associative array (not a sequential) array.

NOTE: This is NOT a great fix. The correct fix would be fixing MWS. This is
merely a bandaid.